Splits out git commands to git wizard
Authored by: Eric Amodio @eamodio
