+++ src/commands.ts
@@ -23,7 +23,7 @@ import './commands/diffWithWorking';
 import './commands/externalDiff';
 import './commands/generateCommitMessage';
 import './commands/ghpr/openOrCreateWorktree';
-import './commands/quickWizard';
+import './commands/gitWizard';
 import './commands/inviteToLiveShare';
 import './commands/inspect';
 import './commands/logging';
@@ -44,6 +44,7 @@ import './commands/openRepoOnRemote';
 import './commands/openRevisionFile';
 import './commands/openWorkingFile';
 import './commands/patches';
+import './commands/quickWizard';
 import './commands/rebaseEditor';
 import './commands/refreshHover';
 import './commands/remoteProviders';
