+++ src/git/actions.ts
@@ -1,13 +1,14 @@
 import type { Uri } from 'vscode';
 import type { BrowseRepoAtRevisionCommandArgs } from '../commands/browseRepoAtRevision';
-import type { QuickWizardCommandArgs, QuickWizardCommandArgsWithCompletion } from '../commands/quickWizard';
+import type { GitWizardCommandArgs } from '../commands/gitWizard';
+import type { QuickWizardCommandArgsWithCompletion } from '../commands/quickWizard.base';
 import { Commands } from '../constants.commands';
 import { defer } from '../system/promise';
 import { executeCommand, executeEditorCommand } from '../system/vscode/command';
 
-export async function executeGitCommand(args: QuickWizardCommandArgs): Promise<void> {
+export async function executeGitCommand(args: GitWizardCommandArgs): Promise<void> {
 	const deferred = defer<void>();
-	void (await executeCommand<QuickWizardCommandArgsWithCompletion>(Commands.GitCommands, {
+	void (await executeCommand<QuickWizardCommandArgsWithCompletion<GitWizardCommandArgs>>(Commands.GitCommands, {
 		...args,
 		completion: deferred,
 	}));
