+++ src/terminal/linkProvider.ts
@@ -1,7 +1,7 @@
 import type { Disposable, TerminalLink, TerminalLinkContext, TerminalLinkProvider } from 'vscode';
 import { commands, window } from 'vscode';
+import type { GitWizardCommandArgs } from '../commands/gitWizard';
 import type { InspectCommandArgs } from '../commands/inspect';
-import type { QuickWizardCommandArgs } from '../commands/quickWizard';
 import type { ShowQuickBranchHistoryCommandArgs } from '../commands/showQuickBranchHistory';
 import type { ShowQuickCommitCommandArgs } from '../commands/showQuickCommit';
 import { Commands } from '../constants.commands';
@@ -62,14 +62,14 @@ export class GitTerminalLinkProvider implements Disposable, TerminalLinkProvider
 			if (match != null) {
 				const [_, git, command] = match;
 
-				const link: GitTerminalLink<QuickWizardCommandArgs> = {
+				const link: GitTerminalLink<GitWizardCommandArgs> = {
 					startIndex: match.index + git.length,
 					length: command.length,
 					tooltip: 'Open in Git Command Palette',
 					command: {
 						command: Commands.GitCommands,
 						args: {
-							command: command as QuickWizardCommandArgs['command'],
+							command: command as GitWizardCommandArgs['command'],
 						},
 					},
 				};
@@ -146,7 +146,7 @@ export class GitTerminalLinkProvider implements Disposable, TerminalLinkProvider
 
 			if (!shaRegex.test(ref)) {
 				if (rangeRegex.test(ref)) {
-					const link: GitTerminalLink<QuickWizardCommandArgs> = {
+					const link: GitTerminalLink<GitWizardCommandArgs> = {
 						startIndex: match.index,
 						length: ref.length,
 						tooltip: 'Show Commits',
